more colorkey goodness

+++ app/editor/portrait_editor/portrait_model.py
@@ -57,26 +57,23 @@ def auto_frame_portrait(portrait: Portrait):
 def auto_colorkey(portrait: Portrait):
     if not portrait.pixmap:
         portrait.pixmap = QPixmap(portrait.full_path)
-    image = portrait.pixmap.toImage()
-    top_left_color = QColor(image.pixel(0, 0)).getRgb()
-    for x in range(image.width()):
-        for y in range(image.height()):
-            if QColor(image.pixel(x, y)).getRgb() == top_left_color:
-                image.setPixelColor(x, y, QColor(*COLORKEY))
-    # since we're messing with data, let's try to be atomic
-    try:
-        shutil.copyfile(portrait.full_path, portrait.full_path + '.bak')
-    except:
-        raise IOError("failed to create backup, aborting auto-colorkey")
-    os.remove(portrait.full_path)
-    try:
-        image.save(portrait.full_path)
-        portrait.pixmap = QPixmap(portrait.full_path)
-        portrait.image = None # reset this so the engine will know to reload
-    except:
-        shutil.move(portrait.full_path + '.bak', portrait.full_path)
-        raise IOError("some file operation failed, aborting auto-colorkey")
-    os.remove(portrait.full_path + '.bak')
+    im = portrait.pixmap.toImage()
+    if im.pixel(0, 0) != editor_utilities.qCOLORKEY:
+        im = editor_utilities.color_convert(im, {im.pixel(0, 0): editor_utilities.qCOLORKEY})
+        # since we're messing with data, let's try to be atomic
+        try:
+            shutil.copyfile(portrait.full_path, portrait.full_path + '.bak')
+        except:
+            raise IOError("failed to create backup, aborting auto-colorkey")
+        os.remove(portrait.full_path)
+        try:
+            im.save(portrait.full_path)
+            portrait.pixmap = QPixmap(portrait.full_path)
+            portrait.image = None # reset this so the engine will know to reload
+        except:
+            shutil.move(portrait.full_path + '.bak', portrait.full_path)
+            raise IOError("some file operation failed, aborting auto-colorkey")
+        os.remove(portrait.full_path + '.bak')
 
 class PortraitModel(ResourceCollectionModel):
     def data(self, index, role):
@@ -113,8 +110,8 @@ class PortraitModel(ResourceCollectionModel):
                     nid = str_utils.get_next_name(nid, [d.nid for d in RESOURCES.portraits])
                     if pix.width() == 128 and pix.height() == 112:
                         # Swap to use colorkey color if it's not
-                        pix = editor_utilities.convert_to_correct_colorkey(pix)
                         new_portrait = Portrait(nid, fn, pix)
+                        auto_colorkey(new_portrait)
                         auto_frame_portrait(new_portrait)
                         RESOURCES.portraits.append(new_portrait)
                     else:
